:recycle: move ITEOp, LogicalOp, and SEq to dedicated modules
