+++ src/Grisette/Core/Data/Class/SEq.hs
@@ -20,14 +20,10 @@
 -- Maintainer  :   siruilu@cs.washington.edu
 -- Stability   :   Experimental
 -- Portability :   GHC only
-module Grisette.Core.Data.Class.Bool
+module Grisette.Core.Data.Class.SEq
   ( -- * Symbolic equality
     SEq (..),
     SEq' (..),
-
-    -- * Symbolic Boolean operations
-    LogicalOp (..),
-    ITEOp (..),
   )
 where
 
@@ -58,20 +54,9 @@ import Generics.Deriving
     type (:+:) (L1, R1),
   )
 import Grisette.Core.Data.BV (IntN, SomeIntN, SomeWordN, WordN)
+import Grisette.Core.Data.Class.LogicalOp (LogicalOp (nots, (&&~)))
 import Grisette.Core.Data.Class.Solvable (Solvable (con))
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
-  ( LinkedRep,
-    SupportedPrim,
-  )
-import Grisette.IR.SymPrim.Data.Prim.PartialEval.Bool
-  ( pevalAndTerm,
-    pevalEqvTerm,
-    pevalITETerm,
-    pevalImplyTerm,
-    pevalNotTerm,
-    pevalOrTerm,
-    pevalXorTerm,
-  )
+import Grisette.IR.SymPrim.Data.Prim.PartialEval.Bool (pevalEqvTerm)
 import {-# SOURCE #-} Grisette.IR.SymPrim.Data.SymPrim
   ( SomeSymIntN (SomeSymIntN),
     SomeSymWordN (SomeSymWordN),
@@ -79,10 +64,6 @@ import {-# SOURCE #-} Grisette.IR.SymPrim.Data.SymPrim
     SymIntN (SymIntN),
     SymInteger (SymInteger),
     SymWordN (SymWordN),
-    binSomeSymIntNR1,
-    binSomeSymWordNR1,
-    type (-~>) (SymGeneralFun),
-    type (=~>) (SymTabularFun),
   )
 
 -- $setup
@@ -126,84 +107,6 @@ class SEq a where
   {-# INLINE (/=~) #-}
   infix 4 /=~
   {-# MINIMAL (==~) | (/=~) #-}
-
--- | Symbolic logical operators for symbolic booleans.
---
--- >>> let t = con True :: SymBool
--- >>> let f = con False :: SymBool
--- >>> let a = "a" :: SymBool
--- >>> let b = "b" :: SymBool
--- >>> t ||~ f
--- true
--- >>> a ||~ t
--- true
--- >>> a ||~ f
--- a
--- >>> a ||~ b
--- (|| a b)
--- >>> t &&~ f
--- false
--- >>> a &&~ t
--- a
--- >>> a &&~ f
--- false
--- >>> a &&~ b
--- (&& a b)
--- >>> nots t
--- false
--- >>> nots f
--- true
--- >>> nots a
--- (! a)
--- >>> t `xors` f
--- true
--- >>> t `xors` t
--- false
--- >>> a `xors` t
--- (! a)
--- >>> a `xors` f
--- a
--- >>> a `xors` b
--- (|| (&& (! a) b) (&& a (! b)))
-class LogicalOp b where
-  -- | Symbolic disjunction
-  (||~) :: b -> b -> b
-  a ||~ b = nots $ nots a &&~ nots b
-  {-# INLINE (||~) #-}
-
-  infixr 2 ||~
-
-  -- | Symbolic conjunction
-  (&&~) :: b -> b -> b
-  a &&~ b = nots $ nots a ||~ nots b
-  {-# INLINE (&&~) #-}
-
-  infixr 3 &&~
-
-  -- | Symbolic negation
-  nots :: b -> b
-
-  -- | Symbolic exclusive disjunction
-  xors :: b -> b -> b
-  a `xors` b = (a &&~ nots b) ||~ (nots a &&~ b)
-  {-# INLINE xors #-}
-
-  -- | Symbolic implication
-  implies :: b -> b -> b
-  a `implies` b = nots a ||~ b
-  {-# INLINE implies #-}
-
-  {-# MINIMAL (||~), nots | (&&~), nots #-}
-
--- | ITE operator for solvable (see "Grisette.Core#solvable")s, including symbolic boolean, integer, etc.
---
--- >>> let a = "a" :: SymBool
--- >>> let b = "b" :: SymBool
--- >>> let c = "c" :: SymBool
--- >>> ites a b c
--- (ite a b c)
-class ITEOp v where
-  ites :: SymBool -> v -> v -> v
 
 -- SEq instances
 #define CONCRETE_SEQ(type) \
@@ -358,54 +261,6 @@ SEQ_BV(SymWordN)
 SEQ_BV_SOME(SomeSymIntN, SymIntN)
 SEQ_BV_SOME(SomeSymWordN, SymWordN)
 #endif
-
--- LogicalOp instances
-instance LogicalOp Bool where
-  (||~) = (||)
-  {-# INLINE (||~) #-}
-  (&&~) = (&&)
-  {-# INLINE (&&~) #-}
-  nots = not
-  {-# INLINE nots #-}
-
--- ITEOp instances
-#define ITEOP_SIMPLE(type) \
-instance ITEOp type where \
-  ites (SymBool c) (type t) (type f) = type $ pevalITETerm c t f; \
-  {-# INLINE ites #-}
-
-#define ITEOP_BV(type) \
-instance (KnownNat n, 1 <= n) => ITEOp (type n) where \
-  ites (SymBool c) (type t) (type f) = type $ pevalITETerm c t f; \
-  {-# INLINE ites #-}
-
-#define ITEOP_BV_SOME(symtype, bf) \
-instance ITEOp symtype where \
-  ites c = bf (ites c) "ites"; \
-  {-# INLINE ites #-}
-
-#define ITEOP_FUN(op, cons) \
-instance (SupportedPrim ca, SupportedPrim cb, LinkedRep ca sa, LinkedRep cb sb) => ITEOp (sa op sb) where \
-  ites (SymBool c) (cons t) (cons f) = cons $ pevalITETerm c t f; \
-  {-# INLINE ites #-}
-
-#if 1
-ITEOP_SIMPLE(SymBool)
-ITEOP_SIMPLE(SymInteger)
-ITEOP_BV(SymIntN)
-ITEOP_BV(SymWordN)
-ITEOP_BV_SOME(SomeSymIntN, binSomeSymIntNR1)
-ITEOP_BV_SOME(SomeSymWordN, binSomeSymWordNR1)
-ITEOP_FUN(=~>, SymTabularFun)
-ITEOP_FUN(-~>, SymGeneralFun)
-#endif
-
-instance LogicalOp SymBool where
-  (SymBool l) ||~ (SymBool r) = SymBool $ pevalOrTerm l r
-  (SymBool l) &&~ (SymBool r) = SymBool $ pevalAndTerm l r
-  nots (SymBool v) = SymBool $ pevalNotTerm v
-  (SymBool l) `xors` (SymBool r) = SymBool $ pevalXorTerm l r
-  (SymBool l) `implies` (SymBool r) = SymBool $ pevalImplyTerm l r
 
 -- | Auxiliary class for 'SEq' instance derivation
 class SEq' f where
