+++ src/Grisette/Lib/Data/List.hs
@@ -20,9 +20,9 @@ where
 import Control.Exception (ArrayException (IndexOutOfBounds))
 import Control.Monad.Except (MonadError (throwError))
 import Grisette.Core.Control.Monad.Union (MonadUnion)
-import Grisette.Core.Data.Class.Bool (SEq ((==~)))
 import Grisette.Core.Data.Class.Error (TransformError (transformError))
 import Grisette.Core.Data.Class.Mergeable (Mergeable)
+import Grisette.Core.Data.Class.SEq (SEq ((==~)))
 import Grisette.Core.Data.Class.SOrd (SOrd ((<=~)))
 import Grisette.Core.Data.Class.SimpleMergeable (mrgIf)
 import Grisette.IR.SymPrim.Data.SymPrim (SymBool, SymInteger)
