+++ test/Grisette/Core/Data/Class/SubstituteSymTests.hs
@@ -19,7 +19,7 @@ import Data.Functor.Sum (Sum (InL, InR))
 import Data.Int (Int16, Int32, Int64, Int8)
 import Data.Word (Word16, Word32, Word64, Word8)
 import GHC.Stack (HasCallStack)
-import Grisette.Core.Data.Class.Bool (LogicalOp ((||~)))
+import Grisette.Core.Data.Class.LogicalOp (LogicalOp ((||~)))
 import Grisette.Core.Data.Class.Substitute (SubstituteSym (substituteSym))
 import Grisette.Core.Data.Class.TestValues (ssymBool, ssymbolBool)
 import Grisette.IR.SymPrim.Data.SymPrim (SymBool)
