+++ test/Grisette/Lib/Control/Monad/ExceptTests.hs
@@ -4,7 +4,7 @@ module Grisette.Lib.Control.Monad.ExceptTests (monadExceptFunctionTests) where
 
 import Control.Monad.Trans.Except (ExceptT (ExceptT), runExceptT)
 import Grisette.Core.Control.Monad.UnionM (UnionM)
-import Grisette.Core.Data.Class.Bool (ITEOp (ites))
+import Grisette.Core.Data.Class.ITEOp (ITEOp (ites))
 import Grisette.Core.Data.Class.SimpleMergeable
   ( UnionLike (unionIf),
     mrgSingle,
